+++ CheForest/src/main/java/com/simplecoding/cheforest/jpa/config/WebSocketConfig.java
@@ -5,14 +5,14 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.security.config.annotation.web.messaging.MessageSecurityMetadataSourceRegistry;
+import org.springframework.security.config.annotation.web.socket.AbstractSecurityWebSocketMessageBrokerConfigurer;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 
 @Configuration
 @EnableWebSocketMessageBroker
 @RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
+public class WebSocketConfig extends AbstractSecurityWebSocketMessageBrokerConfigurer {
 
     private final StompHandler stompHandler;
 
@@ -21,9 +21,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
         // 클라이언트가 WebSocket 연결을 시작할 EndPoint를 설정합니다.
         // Handshake 과정에서 HTTP 세션 정보를 WebSocket 세션으로 안전하게 복사합니다.
         // 이 부분이 JSP의 new SockJS("/ws")와 일치해야 합니다.
-        registry.addEndpoint("/ws")
-                .setAllowedOriginPatterns("*") // 👈 모든 오리진 허용 (개발 시)
-                .withSockJS();
+        registry.addEndpoint("/ws").withSockJS();
     }
 
     @Override
@@ -41,8 +39,31 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
      * 이 핸들러가 Spring Security의 보안 검사보다 먼저 실행되어
      * CONNECT 메시지에 사용자 인증 정보를 올바르게 설정하는 핵심 역할을 합니다.
      */
+    @Override
     public void customizeClientInboundChannel(ChannelRegistration registration) {
         registration.interceptors(stompHandler);
+    }
+
+    /**
+     * WebSocket 메시지에 대한 보안 규칙을 설정합니다.
+     * StompHandler가 먼저 실행된 후 이 규칙이 적용됩니다.
+     */
+    @Override
+    protected void configureInbound(MessageSecurityMetadataSourceRegistry messages) {
+        messages
+                // "/pub/message" 경로로 들어오는 메시지는 인증된 사용자만 허용합니다.
+                .simpDestMatchers("/pub/message").authenticated()
+                // 그 외 다른 모든 메시지(CONNECT, SUBSCRIBE 등)는 인증된 사용자라면 모두 허용합니다.
+                .anyMessage().authenticated();
+    }
+
+    /**
+     * STOMP는 웹소켓을 사용하므로 일반적인 웹 CSRF 토큰을 사용하기 어렵습니다.
+     * 비활성화하는 것이 일반적인 설정입니다.
+     */
+    @Override
+    protected boolean sameOriginDisabled() {
+        return true;
     }
 }
